- Updating branching command to work with the new api of merge. - Adding relevant tests - version up (Breaking change)
